requirements.txt
fastapi
uvicorn
django
requests
python-multipart
mysql-connector-python
Pillow
colorama
